+++ rusty-machine/src/learning/glm.rs
@@ -45,6 +45,7 @@ use learning::SupModel;
 /// The model is generic over a Criterion
 /// which specifies the distribution family and
 /// the link function.
+#[derive(Debug)]
 pub struct GenLinearModel<C: Criterion> {
     parameters: Option<Vector<f64>>,
     criterion: C,
@@ -220,6 +221,7 @@ pub trait LinkFunc {
 /// The Logit link function.
 ///
 /// Used primarily as the canonical link in Binomial Regression.
+#[derive(Clone, Copy, Debug)]
 pub struct Logit;
 
 /// The Logit link function.
@@ -243,6 +245,7 @@ impl LinkFunc for Logit {
 /// The log link function.
 ///
 /// Used primarily as the canonical link in Poisson Regression.
+#[derive(Clone, Copy, Debug)]
 pub struct Log;
 
 /// The log link function.
@@ -265,6 +268,7 @@ impl LinkFunc for Log {
 /// The Identity link function.
 ///
 /// Used primarily as the canonical link in Linear Regression.
+#[derive(Clone, Copy, Debug)]
 pub struct Identity;
 
 /// The Identity link function.
@@ -287,6 +291,7 @@ impl LinkFunc for Identity {
 /// The Bernoulli regression family.
 ///
 /// This is equivalent to logistic regression.
+#[derive(Clone, Copy, Debug)]
 pub struct Bernoulli;
 
 impl Criterion for Bernoulli {
@@ -354,6 +359,7 @@ impl Criterion for Bernoulli {
 }
 
 /// The Binomial regression family.
+#[derive(Debug)]
 pub struct Binomial {
     weights: Vec<f64>,
 }
@@ -426,6 +432,7 @@ impl Criterion for Binomial {
 /// The Normal regression family.
 ///
 /// This is equivalent to the Linear Regression model.
+#[derive(Clone, Copy, Debug)]
 pub struct Normal;
 
 impl Criterion for Normal {
@@ -437,6 +444,7 @@ impl Criterion for Normal {
 }
 
 /// The Poisson regression family.
+#[derive(Clone, Copy, Debug)]
 pub struct Poisson;
 
 impl Criterion for Poisson {