+++ rusty-machine/src/learning/toolkit/activ_fn.rs
@@ -21,6 +21,7 @@ pub trait ActivationFunc {
 }
 
 /// Sigmoid activation function.
+#[derive(Clone, Copy, Debug)]
 pub struct Sigmoid;
 
 impl ActivationFunc for Sigmoid {
@@ -44,6 +45,7 @@ impl ActivationFunc for Sigmoid {
 }
 
 /// Linear activation function.
+#[derive(Clone, Copy, Debug)]
 pub struct Linear;
 
 impl ActivationFunc for Linear {
@@ -61,6 +63,7 @@ impl ActivationFunc for Linear {
 }
 
 /// Exponential activation function.
+#[derive(Clone, Copy, Debug)]
 pub struct Exp;
 
 impl ActivationFunc for Exp {