+++ rusty-machine/src/linalg/matrix/mod.rs
@@ -68,8 +68,8 @@ impl<T> Matrix<T> {
         &self.data
     }
 
-    /// Returns a mutable reference to the underlying data.
-    pub fn mut_data(&mut self) -> &mut Vec<T> {
+    /// Returns a mutable slice of the underlying data.
+    pub fn mut_data(&mut self) -> &mut [T] {
         &mut self.data
     }
 
@@ -363,14 +363,12 @@ impl<T: Copy> Matrix<T> {
     ///
     /// assert_eq!(*b.data(), vec![2.0; 4]);
     /// ```
-    pub fn apply(self, f: &Fn(T) -> T) -> Matrix<T> {
-        let new_data = self.data.into_iter().map(f).collect();
-
-        Matrix {
-            rows: self.rows,
-            cols: self.cols,
-            data: new_data,
+    pub fn apply(mut self, f: &Fn(T) -> T) -> Matrix<T> {
+        for val in self.data.iter_mut() {
+            *val = f(*val);
         }
+
+        self
     }
 }
 
@@ -967,7 +965,7 @@ impl<T: Copy + One + Zero + Mul<T, Output = T>> Mul<T> for Matrix<T> {
     type Output = Matrix<T>;
 
     fn mul(self, f: T) -> Matrix<T> {
-        (&self) * (&f)
+        self * &f
     }
 }
 
@@ -975,8 +973,12 @@ impl<T: Copy + One + Zero + Mul<T, Output = T>> Mul<T> for Matrix<T> {
 impl<'a, T: Copy + One + Zero + Mul<T, Output = T>> Mul<&'a T> for Matrix<T> {
     type Output = Matrix<T>;
 
-    fn mul(self, f: &T) -> Matrix<T> {
-        (&self) * f
+    fn mul(mut self, f: &T) -> Matrix<T> {
+        for val in self.data.iter_mut() {
+            *val = *val * *f;
+        }
+
+        self
     }
 }
 
@@ -1111,7 +1113,7 @@ impl<T: Copy + One + Zero + Add<T, Output = T>> Add<T> for Matrix<T> {
     type Output = Matrix<T>;
 
     fn add(self, f: T) -> Matrix<T> {
-        (&self) + (&f)
+        self + &f
     }
 }
 
@@ -1128,8 +1130,12 @@ impl<'a, T: Copy + One + Zero + Add<T, Output = T>> Add<T> for &'a Matrix<T> {
 impl<'a, T: Copy + One + Zero + Add<T, Output = T>> Add<&'a T> for Matrix<T> {
     type Output = Matrix<T>;
 
-    fn add(self, f: &T) -> Matrix<T> {
-        (&self) + f
+    fn add(mut self, f: &T) -> Matrix<T> {
+        for val in self.data.iter_mut() {
+            *val = *val + *f;
+        }
+
+        self
     }
 }
 
@@ -1152,7 +1158,7 @@ impl<T: Copy + One + Zero + Add<T, Output = T>> Add<Matrix<T>> for Matrix<T> {
     type Output = Matrix<T>;
 
     fn add(self, f: Matrix<T>) -> Matrix<T> {
-        (&self) + (&f)
+        self + &f
     }
 }
 
@@ -1161,7 +1167,7 @@ impl<'a, T: Copy + One + Zero + Add<T, Output = T>> Add<Matrix<T>> for &'a Matri
     type Output = Matrix<T>;
 
     fn add(self, f: Matrix<T>) -> Matrix<T> {
-        self + (&f)
+        f + self
     }
 }
 
@@ -1169,8 +1175,9 @@ impl<'a, T: Copy + One + Zero + Add<T, Output = T>> Add<Matrix<T>> for &'a Matri
 impl<'a, T: Copy + One + Zero + Add<T, Output = T>> Add<&'a Matrix<T>> for Matrix<T> {
     type Output = Matrix<T>;
 
-    fn add(self, f: &Matrix<T>) -> Matrix<T> {
-        (&self) + f
+    fn add(mut self, f: &Matrix<T>) -> Matrix<T> {
+        utils::in_place_vec_bin_op(&mut self.data, &f.data, |x,&y| {*x = *x + y});
+        self
     }
 }
 
@@ -1197,7 +1204,7 @@ impl<T: Copy + One + Zero + Sub<T, Output = T>> Sub<T> for Matrix<T> {
     type Output = Matrix<T>;
 
     fn sub(self, f: T) -> Matrix<T> {
-        (&self) - (&f)
+        self - (&f)
     }
 }
 
@@ -1205,8 +1212,12 @@ impl<T: Copy + One + Zero + Sub<T, Output = T>> Sub<T> for Matrix<T> {
 impl<'a, T: Copy + One + Zero + Sub<T, Output = T>> Sub<&'a T> for Matrix<T> {
     type Output = Matrix<T>;
 
-    fn sub(self, f: &T) -> Matrix<T> {
-        (&self) - f
+    fn sub(mut self, f: &T) -> Matrix<T> {
+        for val in self.data.iter_mut() {
+            *val = *val - *f;
+        }
+
+        self
     }
 }
 
@@ -1238,8 +1249,8 @@ impl<'a, 'b, T: Copy + One + Zero + Sub<T, Output = T>> Sub<&'b T> for &'a Matri
 impl<T: Copy + One + Zero + Sub<T, Output = T>> Sub<Matrix<T>> for Matrix<T> {
     type Output = Matrix<T>;
 
-    fn sub(self, f: Matrix<T>) -> Matrix<T> {
-        (&self) - (&f)
+    fn sub(self, m: Matrix<T>) -> Matrix<T> {
+        self - &m
     }
 }
 
@@ -1247,8 +1258,10 @@ impl<T: Copy + One + Zero + Sub<T, Output = T>> Sub<Matrix<T>> for Matrix<T> {
 impl<'a, T: Copy + One + Zero + Sub<T, Output = T>> Sub<Matrix<T>> for &'a Matrix<T> {
     type Output = Matrix<T>;
 
-    fn sub(self, f: Matrix<T>) -> Matrix<T> {
-        self - (&f)
+    fn sub(self, mut m: Matrix<T>) -> Matrix<T> {
+        utils::in_place_vec_bin_op(&mut m.data, &self.data, |x,&y| {*x = y - *x});
+        
+        m
     }
 }
 
@@ -1256,8 +1269,10 @@ impl<'a, T: Copy + One + Zero + Sub<T, Output = T>> Sub<Matrix<T>> for &'a Matri
 impl<'a, T: Copy + One + Zero + Sub<T, Output = T>> Sub<&'a Matrix<T>> for Matrix<T> {
     type Output = Matrix<T>;
 
-    fn sub(self, f: &Matrix<T>) -> Matrix<T> {
-        (&self) - f
+    fn sub(mut self, m: &Matrix<T>) -> Matrix<T> {
+        utils::in_place_vec_bin_op(&mut self.data, &m.data, |x,&y| {*x = *x - y});
+        
+        self
     }
 }
 
@@ -1284,7 +1299,7 @@ impl<T: Copy + One + Zero + PartialEq + Div<T, Output = T>> Div<T> for Matrix<T>
     type Output = Matrix<T>;
 
     fn div(self, f: T) -> Matrix<T> {
-        (&self) / (&f)
+        self / (&f)
     }
 }
 
@@ -1301,8 +1316,12 @@ impl<'a, T: Copy + One + Zero + PartialEq + Div<T, Output = T>> Div<T> for &'a M
 impl<'a, T: Copy + One + Zero + PartialEq + Div<T, Output = T>> Div<&'a T> for Matrix<T> {
     type Output = Matrix<T>;
 
-    fn div(self, f: &T) -> Matrix<T> {
-        (&self) / f
+    fn div(mut self, f: &T) -> Matrix<T> {
+        for val in self.data.iter_mut() {
+            *val = *val / *f;
+        }
+
+        self
     }
 }
 
@@ -1327,14 +1346,12 @@ impl<'a, 'b, T: Copy + One + Zero + PartialEq + Div<T, Output = T>> Div<&'b T> f
 impl<T: Neg<Output = T> + Copy> Neg for Matrix<T> {
     type Output = Matrix<T>;
 
-    fn neg(self) -> Matrix<T> {
-        let new_data = self.data.iter().map(|v| -*v).collect();
-
-        Matrix {
-            cols: self.cols,
-            rows: self.rows,
-            data: new_data,
+    fn neg(mut self) -> Matrix<T> {
+        for val in self.data.iter_mut() {
+            *val = -*val;
         }
+
+        self
     }
 }