+++ rusty-machine/tests/linalg/vector.rs
@@ -38,6 +38,28 @@ fn vector_f32_mul() {
     let a = Vector::new(vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
     let b = 3.0;
 
+    // Allocating new memory
+    let c = &a * &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], 3.0 * ((i + 1) as f32));
+    }
+
+    // Allocating new memory
+    let c = &a * b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], 3.0 * ((i + 1) as f32));
+    }
+
+    // Reusing memory
+    let c = a.clone() * &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], 3.0 * ((i + 1) as f32));
+    }
+
+    // Reusing memory
     let c = a * b;
 
     for i in 0..6 {
@@ -50,6 +72,28 @@ fn vector_f32_div() {
     let a = Vector::new(vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
     let b = 3.0;
 
+    // Allocating new memory
+    let c = &a / &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((i + 1) as f32) / 3.0);
+    }
+
+    // Allocating new memory
+    let c = &a / b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((i + 1) as f32) / 3.0);
+    }
+
+    // Reusing memory
+    let c = a.clone() / &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((i + 1) as f32) / 3.0);
+    }
+
+    // Reusing memory
     let c = a / b;
 
     for i in 0..6 {
@@ -62,6 +106,28 @@ fn vector_add() {
     let a = Vector::new(vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
     let b = Vector::new(vec![2.0, 3.0, 4.0, 5.0, 6.0, 7.0]);
 
+    // Allocating new memory
+    let c = &a + &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((2 * i + 3) as f32));
+    }
+
+    // Reusing memory
+    let c = &a + b.clone();
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((2 * i + 3) as f32));
+    }
+
+    // Reusing memory
+    let c = a.clone() + &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((2 * i + 3) as f32));
+    }
+
+    // Reusing memory
     let c = a + b;
 
     for i in 0..6 {
@@ -74,6 +140,28 @@ fn vector_f32_add() {
     let a = Vector::new(vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
     let b = 2.0;
 
+    // Allocating new memory
+    let c = &a + &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((i + 1) as f32) + 2.0);
+    }
+
+    // Allocating new memory
+    let c = &a + b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((i + 1) as f32) + 2.0);
+    }
+
+    // Reusing memory
+    let c = a.clone() + &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((i + 1) as f32) + 2.0);
+    }
+
+    // Reusing memory
     let c = a + b;
 
     for i in 0..6 {
@@ -86,6 +174,28 @@ fn vector_sub() {
     let a = Vector::new(vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
     let b = Vector::new(vec![2.0, 3.0, 4.0, 5.0, 6.0, 7.0]);
 
+    // Allocating new memory
+    let c = &a - &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], -1.0);
+    }
+
+    // Reusing memory
+    let c = &a - b.clone();
+
+    for i in 0..6 {
+        assert_eq!(c[i], -1.0);
+    }
+
+    // Reusing memory
+    let c = a.clone() - &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], -1.0);
+    }
+
+    // Reusing memory
     let c = a - b;
 
     for i in 0..6 {
@@ -98,6 +208,28 @@ fn vector_f32_sub() {
     let a = Vector::new(vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
     let b = 2.0;
 
+    // Allocating new memory
+    let c = &a - &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((i + 1) as f32) - 2.0);
+    }
+
+    // Allocating new memory
+    let c = &a - b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((i + 1) as f32) - 2.0);
+    }
+
+    // Reusing memory
+    let c = a.clone() - &b;
+
+    for i in 0..6 {
+        assert_eq!(c[i], ((i + 1) as f32) - 2.0);
+    }
+
+    // Reusing memory
     let c = a - b;
 
     for i in 0..6 {
@@ -112,6 +244,4 @@ fn vector_norm() {
     let b = a.norm();
 
     assert_eq!(b, (1. + 4. + 9. + 16. + 25. + 36. as f32).sqrt());
-
-
 }